refactor: remove shared directory
