+++ packages/core/src/handler.ts
@@ -4,7 +4,19 @@ import {
   ListenerMaps,
   DrawListenerType,
 } from './types'
-import { getPassiveOptions } from './shared/getPassiveOptions'
+
+export const getPassiveOptions = (
+  passive = true
+): boolean | { passive: boolean } => {
+  try {
+    const check = () => null
+    window.addEventListener('testPassive', check, { passive })
+    window.removeEventListener('testPassive', check)
+    return { passive }
+  } catch (e) {
+    return false
+  }
+}
 
 const listenerMaps: ListenerMaps = {
   pointer: {
