+++ packages/core/src/index.ts
@@ -3,6 +3,6 @@ export * from './convert'
 export * from './renderer'
 export * from './svg'
 export * from './throttle'
-export * from './shared'
+export * from './utils'
 export * from './handler'
 export * from './types'
