--- packages/core/src/shared/camel2kebab.ts
@@ -1,2 +1,0 @@
-export const camel2kebab = (str: string): string =>
-  str.replace(/([a-z])([A-Z])/g, '$1-$2').toLowerCase()
