--- packages/core/src/shared/index.ts
@@ -1,6 +1,0 @@
-export * from './camel2kebab'
-export * from './createSvgElement'
-export * from './download'
-export * from './kebab2camel'
-export * from './roundUp'
-export * from './svg2base64'
