--- packages/core/src/shared/isAlmostSame.ts
@@ -1,2 +1,0 @@
-export const isAlmostSameNumber = (a: number, b: number) =>
-  Math.trunc(a) === Math.trunc(b)
