--- packages/core/src/shared/kebab2camel.ts
@@ -1,2 +1,0 @@
-export const kebab2camel = (str: string): string =>
-  str.replace(/-([a-z])/g, (a: string, b: string) => b.toUpperCase())
