--- packages/core/src/shared/roundUp.ts
@@ -1,1 +1,0 @@
-export const roundUp = (num: number, digits = 2): number => +num.toFixed(digits)
