--- packages/core/src/shared/svg2base64.ts
@@ -1,2 +1,0 @@
-export const svg2base64 = (svg: string): string =>
-  `data:image/svg+xml;base64,${btoa(svg)}`
